[Append] 文件上传模块开发，当前开发将data_info与upload_task分离，存在问题。

+++ common/api/src/main/java/com/urban/carbon/api/upload/response/data/UploadChunkInfo.java
@@ -24,7 +24,7 @@ public class UploadChunkInfo implements Serializable {
     /**
      * 分片索引
      */
-    private Long chunkIndex;
+    private Integer chunkIndex;
 
     /**
      * 是否接收成功
@@ -42,7 +42,7 @@ public class UploadChunkInfo implements Serializable {
      * @param chunkIndex 分片索引
      * @param received 是否接收成功
      */
-    public void extracted(String fileId, Long chunkIndex, boolean received) {
+    public void extracted(String fileId, Integer chunkIndex, boolean received) {
         this.setFileId(fileId);
         this.setChunkIndex(chunkIndex);
         this.setReceived(received);
